# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kovex ERP API - Smart Business Management for SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List all purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchase-orders]
      summary: Mark purchase order as received (increases stock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # Purchase Invoices
  /purchase-invoices:
    get:
      operationId: listPurchaseInvoices
      tags: [purchase-invoices]
      summary: List all purchase invoices
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of purchase invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseInvoice"
    post:
      operationId: createPurchaseInvoice
      tags: [purchase-invoices]
      summary: Create a purchase invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"
  /purchase-invoices/{id}:
    get:
      operationId: getPurchaseInvoice
      tags: [purchase-invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"
    patch:
      operationId: updatePurchaseInvoice
      tags: [purchase-invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"

  # Users
...
components:
  schemas:
...
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitCost:
          type: number
        total:
          type: number
      required: [id, productId, productName, quantity, unitCost, total]
    PurchaseOrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitCost:
          type: number
      required: [productId, quantity, unitCost]
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        expectedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        createdAt:
          type: string
      required:
        [
          id,
          reference,
          supplierId,
          supplierName,
          status,
          totalAmount,
          createdAt,
          items,
        ]
    PurchaseOrderInput:
      type: object
      properties:
        supplierId:
          type: integer
        expectedDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
      required: [supplierId, items]
    PurchaseOrderUpdate:
      type: object
      properties:
        status:
          type: string
        expectedDate:
          type: string
        notes:
          type: string

    # Purchase Invoice
    PurchaseInvoice:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        purchaseOrderId:
          type: ["integer", "null"]
        status:
          type: string
        totalAmount:
          type: number
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        [
          id,
          reference,
          supplierId,
          supplierName,
          status,
          totalAmount,
          createdAt,
        ]
    PurchaseInvoiceInput:
      type: object
      properties:
        supplierId:
          type: integer
        purchaseOrderId:
          type: integer
        totalAmount:
          type: number
        dueDate:
          type: string
        notes:
          type: string
      required: [supplierId, totalAmount]
    PurchaseInvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        dueDate:
          type: string
        notes:
          type: string

    # Users
...